+++ crates/ibc/src/core/context.rs
@@ -1,10 +1,10 @@
+mod acknowledgement;
 mod timeout;
 
 use crate::prelude::*;
 
-use timeout::{timeout_packet_execute, timeout_packet_validate};
-
-use self::timeout::TimeoutMsgType;
+use self::acknowledgement::{acknowledgement_packet_execute, acknowledgement_packet_validate};
+use self::timeout::{timeout_packet_execute, timeout_packet_validate, TimeoutMsgType};
 
 use super::{
     ics02_client::error::ClientError,
@@ -26,14 +26,14 @@ use crate::core::ics04_channel::channel::{ChannelEnd, Counterparty, Order, State
 use crate::core::ics04_channel::commitment::{AcknowledgementCommitment, PacketCommitment};
 use crate::core::ics04_channel::context::calculate_block_delay;
 use crate::core::ics04_channel::events::{
-    AcknowledgePacket, CloseConfirm, CloseInit, OpenAck, OpenConfirm, OpenInit, OpenTry,
-    ReceivePacket, WriteAcknowledgement,
+    CloseConfirm, CloseInit, OpenAck, OpenConfirm, OpenInit, OpenTry, ReceivePacket,
+    WriteAcknowledgement,
 };
 use crate::core::ics04_channel::handler::{
-    acknowledgement, chan_close_confirm, chan_close_init, chan_open_ack, chan_open_confirm,
-    chan_open_init, chan_open_try, recv_packet,
+    chan_close_confirm, chan_close_init, chan_open_ack, chan_open_confirm, chan_open_init,
+    chan_open_try, recv_packet,
 };
-use crate::core::ics04_channel::msgs::acknowledgement::{Acknowledgement, MsgAcknowledgement};
+use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
 use crate::core::ics04_channel::msgs::chan_close_confirm::MsgChannelCloseConfirm;
 use crate::core::ics04_channel::msgs::chan_close_init::MsgChannelCloseInit;
 use crate::core::ics04_channel::msgs::chan_open_ack::MsgChannelOpenAck;
@@ -1266,103 +1266,6 @@ where
 
         for log_message in extras.log {
             ctx_b.log_message(log_message);
-        }
-    }
-
-    Ok(())
-}
-
-fn acknowledgement_packet_validate<ValCtx>(
-    ctx_a: &ValCtx,
-    module_id: ModuleId,
-    msg: MsgAcknowledgement,
-) -> Result<(), ContextError>
-where
-    ValCtx: ValidationContext,
-{
-    acknowledgement::validate(ctx_a, &msg)?;
-
-    let module = ctx_a
-        .get_route(&module_id)
-        .ok_or(ChannelError::RouteNotFound)?;
-
-    module
-        .on_acknowledgement_packet_validate(&msg.packet, &msg.acknowledgement, &msg.signer)
-        .map_err(ContextError::PacketError)
-}
-
-fn acknowledgement_packet_execute<ExecCtx>(
-    ctx_a: &mut ExecCtx,
-    module_id: ModuleId,
-    msg: MsgAcknowledgement,
-) -> Result<(), ContextError>
-where
-    ExecCtx: ExecutionContext,
-{
-    let port_chan_id_on_a = (msg.packet.port_on_a.clone(), msg.packet.chan_on_a.clone());
-    let chan_end_on_a = ctx_a.channel_end(&port_chan_id_on_a)?;
-    let conn_id_on_a = &chan_end_on_a.connection_hops()[0];
-
-    // In all cases, this event is emitted
-    ctx_a.emit_ibc_event(IbcEvent::AcknowledgePacket(AcknowledgePacket::new(
-        msg.packet.clone(),
-        chan_end_on_a.ordering,
-        conn_id_on_a.clone(),
-    )));
-
-    // check if we're in the NO-OP case
-    if ctx_a
-        .get_packet_commitment(&(
-            msg.packet.port_on_a.clone(),
-            msg.packet.chan_on_a.clone(),
-            msg.packet.sequence,
-        ))
-        .is_err()
-    {
-        // This error indicates that the timeout has already been relayed
-        // or there is a misconfigured relayer attempting to prove a timeout
-        // for a packet never sent. Core IBC will treat this error as a no-op in order to
-        // prevent an entire relay transaction from failing and consuming unnecessary fees.
-        return Ok(());
-    };
-
-    let module = ctx_a
-        .get_route_mut(&module_id)
-        .ok_or(ChannelError::RouteNotFound)?;
-
-    let (extras, cb_result) =
-        module.on_acknowledgement_packet_execute(&msg.packet, &msg.acknowledgement, &msg.signer);
-
-    cb_result?;
-
-    // apply state changes
-    {
-        let commitment_path = CommitmentsPath {
-            port_id: msg.packet.port_on_a.clone(),
-            channel_id: msg.packet.chan_on_a.clone(),
-            sequence: msg.packet.sequence,
-        };
-        ctx_a.delete_packet_commitment(commitment_path)?;
-
-        if let Order::Ordered = chan_end_on_a.ordering {
-            // Note: in validation, we verified that `msg.packet.sequence == nextSeqRecv`
-            // (where `nextSeqRecv` is the value in the store)
-            ctx_a.store_next_sequence_ack(port_chan_id_on_a, msg.packet.sequence.increment())?;
-        }
-    }
-
-    // emit events and logs
-    {
-        ctx_a.log_message("success: packet acknowledgement".to_string());
-
-        // Note: Acknowledgement event was emitted at the beginning
-
-        for module_event in extras.events {
-            ctx_a.emit_ibc_event(IbcEvent::AppModule(module_event));
-        }
-
-        for log_message in extras.log {
-            ctx_a.log_message(log_message);
         }
     }
